fix: add all testnet and mainnet contract code
This replaces the rolling version of each CityCoin contract in favor of storing their on-chain equivalents for mainnet and testnet. The goal here is to give a clear picture of what's developed locally or deployed to testnet/mainnet, without the need to follow links or grab additional source code.

From here, new versions of contracts can be iterated on in the local folders, then once deployed a new file should be created in the respective testnet/mainnet folders.