feat: add v2 auth contracts, models, and tests
Note that a number of tests were skipped as they require interaction with a new core contract. Given the size of this PR already, it would be better to build a 2nd PR from here with those changes and anything else that finishes up what's needed in the tracking issue.