fix: separate v1 and v2 contracts to isolate interactions
There was some overlap in the tests where a V1 contract couldn't accept a call because it was expecting V2 authorization and other little nuances that were better fixed here then trying to adapt in the tests.

There are now duplicate core V1 and V2 contracts for each city that can be used for simple upgrade tests, and the scenario for the proposed upgrades can be set up as a separate test.