fix: split base (prev: core) tests
This splits up the base (prev: core) tests the same way the city tests are broken up to prevent OOM in Clarinet on the GitHub Action.

This also renames the "core" tests to be the "base" tests to avoid confusion between a core contract and items that are part of the core protocol (e.g. generic "base" tests).