fix: implement v2 traits and emissions changes
This introduces two new traits that support the same as V1 traits plus new function signatures for requested CCIP updates.

This also makes it so a core contract will always query the token contract for coinbase thresholds and amounts, and the auth contract can update those values in the core contract, triggering the same update.