fix: update activation logic to use target block
Introduces a constant for the original MiamiCoin activation height, and a new variable activationTarget, to separate the activation block height used for coinbase calculations from the target block height to start mining/stacking.

Also removes some repetitive code by replacing (var-get activationReached) with the read-only function (get-activation-status) which returns the same result.

Also adds the (is-activated) function and uses it in place of (get-activation-status) for guarded functions.