fix: bring changes to v1 contracts for testing
This commit can be reverted or the changes removed when done, but will show that both the core-v2 tests still pass after the change, and adds a new test to ensure both CityCoins and STX are transferred when claiming in the current cycle.

This will also provide an artifact of the testing with the related github action.