fix: update tests to handle 0 as activation delay
Most of the tests would assume activation delay equals a number of blocks, then subtract one, creating a situation where subtracting one ends up one block behind the chain and causes errors.

There were also a few tests for mine-tokens and mine-many that checked if the activation target is reached, and those are skipped since the target is the same block as the upgrade.