fix: update and simplify gh workflow
The prior version was split into sections due to an OOM issue with Clarinet. This has since been fixed, so we should be able to run all the tests with a single command again. This has the added benefit of picking up any new tests without having to specify them within the yaml each time.